Uninstall the cherry-providing watcher.
There is no need for installing a watcher to watch the score of the user in order to provide him/her with cherries at appropriate intervals. The same logic can be moved to the place after pacman actually makes a move by "walking".
This will increase the response time of the game due to the reduced number of $digest triggers.